CA-181442: Make the list that contained all NewUploadedPatches a Dictionary that matches the patch with its path. This is helpful when trying to upload to a server where the update it is already uploaded (and therefore we do not do the upload again and we cannot get the patch). Adds a flag in the UploadPatchAction that is used when checking if the file should be deleted on cancel.

+++ XenModel/Actions/Pool_Patch/UploadPatchAction.cs
@@ -50,6 +50,7 @@ namespace XenAdmin.Actions
 
         private readonly IList<Host> retailHosts;
         private readonly string retailPatchPath;
+        private readonly bool deleteFileOnCancel;
 
         /// <summary>
         /// This constructor is used to upload a single 'normal' patch
@@ -57,13 +58,14 @@ namespace XenAdmin.Actions
         /// <param name="connection"></param>
         /// <param name="path"></param>
         public UploadPatchAction(IXenConnection connection, string path)
-            : this(connection, path, false)
+            : this(connection, path, false, false)
         {}
 
-        public UploadPatchAction(IXenConnection connection, string path, bool suppressHistory)
+        public UploadPatchAction(IXenConnection connection, string path, bool suppressHistory, bool deleteFileOnCancel)
             : base(connection, null, Messages.UPLOADING_PATCH, suppressHistory)
         {
             Host master = Helpers.GetMaster(connection);
+            this.deleteFileOnCancel = deleteFileOnCancel;
             if (master == null)
                 throw new NullReferenceException();
 
@@ -228,7 +230,7 @@ namespace XenAdmin.Actions
                 }
                 catch(CancelledException e)
                 {
-                     if(File.Exists(retailPatchPath))
+                     if(deleteFileOnCancel && File.Exists(retailPatchPath))
                      {
                         File.Delete(retailPatchPath);
                      }
